+++ src/linklayer/hybridpon/OltSchedulerSSSF.cc
@@ -3,10 +3,10 @@
 /// @author Kyeong Soo (Joseph) Kim <kyeongsoo.kim@gmail.com>
 /// @date   2010-02-22
 ///
-/// @brief  Implements 'OltSchedulerSSSF' class for a hybrid TDM/WDM-PON OLT.
+/// @brief  Implements OltSchedulerSSSF class for a hybrid TDM/WDM-PON OLT.
 ///
 /// @note
-/// This file implements an 'OltSchedulerSSSF' class for the &quot;Sequential
+/// This file implements an OltSchedulerSSSF class for the &quot;Sequential
 /// Scheduling with	Schedule-time Framing (S<sup>3</sup>F)&quot; algorithm reported in
 /// [1] for SUCCESS-HPON.
 ///
@@ -28,10 +28,12 @@
 ///
 /// @par References:
 /// <ol>
-///	<li>Kyeong Soo Kim, David Gutierrez, Fu-Tai An, and Leonid G. Kazovsky,
-///		&quot;Design and performance analysis of scheduling algorithms for WDM-PON
-///		under SUCCESS-HPON architecture,&quot; IEEE/OSA Journal of Lightwave Technology,
-///		vol. 23, no. 11, pp. 3716-3731, Nov. 2005.
+///	<li>
+/// Kyeong Soo Kim, David Gutierrez, Fu-Tai An, and Leonid G. Kazovsky,
+///	&quot;Design and performance analysis of scheduling algorithms for WDM-PON
+///	under SUCCESS-HPON architecture,&quot; IEEE/OSA Journal of Lightwave Technology,
+///	vol. 23, no. 11, pp. 3716-3731, Nov. 2005.
+/// </li>
 /// </ol>
 ///
 /// @remarks Copyright (C) 2010 Kyeong Soo (Joseph) Kim. All rights reserved.
@@ -995,7 +997,7 @@ void OltSchedulerSSSF::initialize(void)
 	OltScheduler::initialize();
 
 	// initialize OltSchedulerSSSF NED parameters
-	voqSize = par("voqSize");
+	voqSize = par("voqSize").longValue();
 	//	rsDepth = par("rsDepth");
 
 	// initialize VOQ counters
